% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstantDef.R
\name{Define_Variables}
\alias{Define_Variables}
\title{Create and Plot Planck, Wein and Kircjhoff data}
\usage{
Define_Variables(WminC = 0.1, WmaxC = 3000, WintervC = 1)
}
\arguments{
\item{WminC}{Min. Wavelenth}

\item{WmaxC}{Max. Wavelenth}

\item{WintervC}{Interval Wavelenth}
}
\value{
Nothing, create the constant values and the wavelength vector
}
\description{
This functions creates the constant values
}
