% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunction.R
\name{DefinePlanckData}
\alias{DefinePlanckData}
\title{Create and Plot Planck, Wein and Kircjhoff data}
\usage{
DefinePlanckData(
  Tmin = 1000,
  Tmax = 10000,
  Tinterv = 1000,
  Wmin = 0.01,
  Wmax = 3000,
  Winterv = 1
)
}
\arguments{
\item{Tmin}{Min. Temperature}

\item{Tmax}{Max. Temperature}

\item{Tinterv}{Interval Temperature}

\item{Wmin}{Min. Wavelength}

\item{Wmax}{Max. Wavelength}

\item{Winterv}{Interval Wavelength}
}
\value{
Nothing, just create dataframes
}
\description{
This functions tranlate math lenguaje to R datasets
}
