% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gifdata.R
\name{GifPlanckData}
\alias{GifPlanckData}
\title{Say hello to the EAGLE}
\usage{
GifPlanckData(Planck_Graph, Time = 0.01)
}
\arguments{
\item{Planck_Graph}{dataframe with planck data}

\item{Time}{Fade time}
}
\value{
Gif
}
\description{
This function says hello to new EAGLES
}
