% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gifdata.R
\name{Planck_Gif}
\alias{Planck_Gif}
\title{Say hello to the EAGLE}
\usage{
Planck_Gif(Planck_Graph)
}
\arguments{
\item{Planck_Graph}{dataframe with planck data}
}
\value{
Nothing, just say Hello
}
\description{
This function says hello to new EAGLES
}
