% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotpdata.R
\name{Planck_Plot}
\alias{Planck_Plot}
\title{Say hello to the EAGLE}
\usage{
Planck_Plot(PlankData_F)
}
\arguments{
\item{PlankData_F}{dataframe with planck data}
}
\value{
Graph of dataframe!
}
\description{
This function says hello to new EAGLES
}
