% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunction.R
\name{Plank}
\alias{Plank}
\title{Create and Plot Planck, Wein and Kircjhoff data}
\usage{
Plank(Wavelength, Temperature)
}
\arguments{
\item{Tmin}{Min. Temperature}

\item{Tmax}{Max. Temperature}

\item{Tinterv}{Interval Temperature}
}
\value{
PlanckVal

Nothing, just create dataframes
}
\description{
This functions tranlate math lenguaje to R datasets
}
