% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlanckFunction.R
\name{Plank_Rad}
\alias{Plank_Rad}
\title{Create and Plot Planck, Wein and Kircjhoff data}
\usage{
Plank_Rad(Wavelength, Temperature)
}
\arguments{
\item{Wavelength}{Wavelenth}

\item{Temperature}{Temperature}
}
\value{
Radiance value for a given wavelenth and temperature
}
\description{
This functions tranlate math lenguaje to R datasets
}
