% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_Function.R
\name{Variables_Planck}
\alias{Variables_Planck}
\title{Create and Plot Planck, Wein and Kircjhoff data}
\usage{
Variables_Planck()
}
\arguments{
\item{Tmin}{Min. Temperature}

\item{Tmax}{Max. Temperature}

\item{Tinterv}{Interval Temperature}
}
\value{
Nothing, just say Hello
}
\description{
This functions tranlate math lenguaje to R datasets
}
