% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlanckFunction.R
\name{Plank}
\alias{Plank}
\title{Create and Plot Planck, Wein and Kircjhoff data}
\usage{
Plank(Wavelength, Temperature)
}
\arguments{
\item{Wavelength}{Min. Temperature}

\item{Temperature}{Max. Temperature}
}
\value{
PlanckVal
}
\description{
This functions tranlate math lenguaje to R datasets
}
